Separating common utility functions (#8115)
