+++ eng/tools/versioning/VersionUtils.js
@@ -1,65 +1,6 @@
-const fs = require("fs");
 const path = require("path");
-const parse = require("../../../common/lib/jju/parse").parse;
-const { promisify } = require("util");
-
-const readFileAsync = promisify(fs.readFile);
-const writeFileAsync = promisify(fs.writeFile);
+const { readFile, writeFile } = require("eng-package-utils");
 var spawnSync = require("child_process").spawnSync, child;
-
-async function readFile(filename) {
-  return await readFileAsync(filename, { encoding: "utf-8" });
-}
-
-async function writeFile(filename, contents) {
-  return await writeFileAsync(filename, contents);
-}
-
-async function readFileJson(filename) {
-  try {
-    const fileContents = await readFile(filename);
-    const jsonResult = parse(fileContents);
-    return jsonResult;
-  } catch (ex) {
-    console.error(ex);
-  }
-}
-
-async function writePackageJson(filename, contentObject) {
-  try {
-    const contentString = JSON.stringify(contentObject, null, 2);
-    await writeFile(filename, `${contentString}\n`);
-  } catch (ex) {
-    console.error(ex);
-  }
-}
-
-//This gets the list of rush packages as well as their packageJsons
-//This is specifically used in set-dev-dependencies script
-const getRushPackageJsons = async repoRoot => {
-  const rushPath = path.resolve(path.join(repoRoot, "rush.json"));
-  const baseDir = path.dirname(rushPath);
-  const rushJson = parse(await readFile(rushPath, "utf8"));
-  const packageData = {};
-
-  for (const proj of rushJson.projects) {
-    const filePath = path.join(baseDir, proj.projectFolder, "package.json");
-    const packageJson = parse(await readFile(filePath, "utf8"));
-    packageData[packageJson.name] = {
-      src: filePath,
-      json: packageJson,
-      versionPolicy: proj.versionPolicyName,
-      projectFolder: proj.projectFolder,
-      newVer: undefined
-    };
-  }
-  return packageData;
-};
-
-async function getRushSpec(repoRoot) {
-  const rushPath = path.resolve(path.join(repoRoot, "rush.json"));
-  return await readFileJson(rushPath);
-}
 
 // This is done to update files which are only periodically generated and
 // checked in. Since these files could be generated once between many versions
@@ -115,11 +56,5 @@ function updateChangelog(targetPackagePath, repoRoot, newVersion, unreleased, re
 // and adapted to exclude beginning of line (^) and end of line ($) anchors.
 const semverRegex = `(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?`;
 
-module.exports.readFile = readFile;
-module.exports.readFileJson = readFileJson;
-module.exports.writeFile = writeFile;
-module.exports.writePackageJson = writePackageJson;
-module.exports.getRushSpec = getRushSpec;
 module.exports.updatePackageConstants = updatePackageConstants;
-module.exports.getRushPackageJsons = getRushPackageJsons;
 module.exports.updateChangelog = updateChangelog;
