+++ eng/tools/versioning/set-version.js
@@ -25,6 +25,7 @@ let argv = require("yargs")
 
 const path = require("path");
 const versionUtils = require("./VersionUtils");
+const packageUtils = require("eng-package-utils");
 
 async function main(argv) {
   const artifactName = argv["artifact-name"];
@@ -33,7 +34,7 @@ async function main(argv) {
   const dryRun = argv["dry-run"];
 
   const packageName = artifactName.replace("azure-", "@azure/");
-  const rushSpec = await versionUtils.getRushSpec(repoRoot);
+  const rushSpec = await packageUtils.getRushSpec(repoRoot);
 
   const targetPackage = rushSpec.projects.find(
     packageSpec => packageSpec.packageName == packageName
@@ -42,7 +43,7 @@ async function main(argv) {
   const targetPackagePath = path.join(repoRoot, targetPackage.projectFolder);
   const packageJsonLocation = path.join(targetPackagePath, "package.json");
 
-  const packageJsonContents = await versionUtils.readFileJson(
+  const packageJsonContents = await packageUtils.readFileJson(
     packageJsonLocation
   );
 
@@ -58,7 +59,7 @@ async function main(argv) {
     ...packageJsonContents,
     version: newVersion
   };
-  await versionUtils.writePackageJson(packageJsonLocation, updatedPackageJson);
+  await packageUtils.writePackageJson(packageJsonLocation, updatedPackageJson);
 
   await versionUtils.updatePackageConstants(
     targetPackagePath,
